requirements.txt
duetwebapi
pyyaml
colorama